Migrated settings to Ids instead of gallery name

+++ WeddingShare/Controllers/AccountController.cs
@@ -189,14 +189,17 @@ namespace WeddingShare.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(AccountTabs tab = AccountTabs.Reviews)
         {
             if (User?.Identity == null || !User.Identity.IsAuthenticated)
             { 
                 return Redirect("/");
             }
 
-            var model = new IndexModel();
+            var model = new IndexModel()
+            {
+                ActiveTab = tab
+            };
 
             var deviceType = HttpContext.Session.GetString(SessionKey.DeviceType);
             if (string.IsNullOrWhiteSpace(deviceType))
@@ -209,35 +212,32 @@ namespace WeddingShare.Controllers
             {
                 var user = await _database.GetUser(User.Identity.GetUserId());
                 if (user != null)
-                { 
-                    if (!await _settings.GetOrDefault(Settings.Basic.SingleGalleryMode, false))
+                {
+                    if (tab == AccountTabs.Reviews)
                     {
+                        model.PendingRequests = await GetPendingReviews();
+                    }
+                    else if (tab == AccountTabs.Galleries)
+                    { 
                         model.Galleries = await _database.GetAllGalleries();
-                        if (model.Galleries != null)
-                        { 
-                            var all = await _database.GetGallery(0);
-                            if (all != null)
-                            { 
-                                model.Galleries.Add(all);
-                            }
-                        }
-
-                        model.PendingRequests = await _database.GetPendingGalleryItems();
                     }
-                    else
+                    else if (tab == AccountTabs.Users)
                     {
-                        var gallery = await _database.GetGallery("default");
-                        if (gallery != null)
-                        { 
-                            model.Galleries = new List<GalleryModel>() { gallery };
-                            model.PendingRequests = await _database.GetPendingGalleryItems(gallery.Id);
-                        }
+                        model.Users = await _database.GetAllUsers();
                     }
-
-                    model.CustomResources = await _database.GetAllCustomResources();
-                    model.Users = await _database.GetAllUsers();
-                    model.Settings = (await _database.GetAllSettings())?.ToDictionary(x => x.Id.ToUpper(), x => x.Value ?? string.Empty);
-                    model.AuditLogs = await _database.GetAuditLogs();
+                    else if (tab == AccountTabs.Resources)
+                    {
+                        model.CustomResources = await _database.GetAllCustomResources();
+                    }
+                    else if (tab == AccountTabs.Settings)
+                    {
+                        model.Settings = (await _database.GetAllSettings())?.ToDictionary(x => x.Id.ToUpper(), x => x.Value ?? string.Empty);
+                        model.CustomResources = await _database.GetAllCustomResources();
+                    }
+                    else if (tab == AccountTabs.Audit)
+                    {
+                        model.AuditLogs = await _database.GetAuditLogs();
+                    }
                 }
             }
             catch (Exception ex)
@@ -278,7 +278,7 @@ namespace WeddingShare.Controllers
                     }
                     else
                     {
-                        var gallery = await _database.GetGallery("default");
+                        var gallery = await _database.GetGallery(1);
                         if (gallery != null)
                         {
                             result = new List<GalleryModel>() { gallery };
@@ -298,30 +298,19 @@ namespace WeddingShare.Controllers
         [RequiresRole(Permission = AccessPermissions.Review_View)]
         public async Task<IActionResult> PendingReviews()
         {
+            var galleries = new List<PhotoGallery>();
+
             if (User?.Identity == null || !User.Identity.IsAuthenticated)
             {
                 return Redirect("/");
             }
-
-            List<GalleryItemModel>? result = null;
 
             try
             {
                 var user = await _database.GetUser(User.Identity.GetUserId());
                 if (user != null)
                 {
-                    if (!await _settings.GetOrDefault(Settings.Basic.SingleGalleryMode, false))
-                    {
-                        result = await _database.GetPendingGalleryItems();
-                    }
-                    else
-                    {
-                        var gallery = await _database.GetGallery("default");
-                        if (gallery != null)
-                        {
-                            result = await _database.GetPendingGalleryItems(gallery.Id);
-                        }
-                    }
+                    galleries = await GetPendingReviews();
                 }
             }
             catch (Exception ex)
@@ -329,7 +318,7 @@ namespace WeddingShare.Controllers
                 _logger.LogError(ex, $"{_localizer["Pending_Uploads_Failed"].Value} - {ex?.Message}");
             }
 
-            return PartialView("~/Views/Account/Partials/PendingReviews.cshtml", result ?? new List<GalleryItemModel>());
+            return PartialView("~/Views/Account/Partials/PendingReviews.cshtml", galleries);
         }
 
         [HttpGet]
@@ -447,7 +436,7 @@ namespace WeddingShare.Controllers
                 var gallery = await _database.GetGallery(galleryId);
                 if (!string.IsNullOrWhiteSpace(gallery?.Name))
                 { 
-                    model.Settings = (await _database.GetAllSettings(gallery.Name))?.Where(x => x.Id.StartsWith(Settings.Gallery.BaseKey, StringComparison.OrdinalIgnoreCase))?.ToDictionary(x => x.Id.ToUpper(), x => x.Value ?? string.Empty);
+                    model.Settings = (await _database.GetAllSettings(gallery.Id))?.Where(x => x.Id.StartsWith(Settings.Gallery.BaseKey, StringComparison.OrdinalIgnoreCase))?.ToDictionary(x => x.Id.ToUpper(), x => x.Value ?? string.Empty);
                     model.CustomResources = await _database.GetAllCustomResources();
                 }
             }
@@ -469,42 +458,46 @@ namespace WeddingShare.Controllers
                 {
                     var review = await _database.GetPendingGalleryItem(id);
                     if (review != null)
-                    { 
-                        var galleryDir = Path.Combine(UploadsDirectory, review.GalleryName);
-                        var reviewFile = Path.Combine(galleryDir, "Pending", review.Title);
-                        if (action == ReviewAction.APPROVED)
-                        {
-                            _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(galleryDir, review.Title));
-
-                            review.State = GalleryItemState.Approved;
-                            await _database.EditGalleryItem(review);
-
-                            await _audit.LogAction(User?.Identity?.Name, $"'{review.Title}' {_localizer["Audit_ItemApprovedInGallery"].Value} '{review.GalleryName}'");
-                        }
-                        else if (action == ReviewAction.REJECTED)
-                        {
-                            var retain = await _settings.GetOrDefault(Settings.Gallery.RetainRejectedItems, false);
-                            if (retain)
+                    {
+                        var gallery = await _database.GetGallery(review.GalleryId);
+                        if (gallery != null)
+                        { 
+                            var galleryDir = Path.Combine(UploadsDirectory, gallery.Name);
+                            var reviewFile = Path.Combine(galleryDir, "Pending", review.Title);
+                            if (action == ReviewAction.APPROVED)
                             {
-                                var rejectedDir = Path.Combine(galleryDir, "Rejected");
-                                _fileHelper.CreateDirectoryIfNotExists(rejectedDir);
-                                _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(rejectedDir, review.Title));
+                                _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(galleryDir, review.Title));
+
+                                review.State = GalleryItemState.Approved;
+                                await _database.EditGalleryItem(review);
+
+                                await _audit.LogAction(User?.Identity?.Name, $"'{review.Title}' {_localizer["Audit_ItemApprovedInGallery"].Value} '{gallery.Name}'");
                             }
-                            else
+                            else if (action == ReviewAction.REJECTED)
                             {
-                                _fileHelper.DeleteFileIfExists(reviewFile);
+                                var retain = await _settings.GetOrDefault(Settings.Gallery.RetainRejectedItems, false);
+                                if (retain)
+                                {
+                                    var rejectedDir = Path.Combine(galleryDir, "Rejected");
+                                    _fileHelper.CreateDirectoryIfNotExists(rejectedDir);
+                                    _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(rejectedDir, review.Title));
+                                }
+                                else
+                                {
+                                    _fileHelper.DeleteFileIfExists(reviewFile);
+                                }
+
+                                await _database.DeleteGalleryItem(review);
+
+                                await _audit.LogAction(User?.Identity?.Name, $"'{review.Title}' {_localizer["Audit_ItemRejectedInGallery"].Value} '{gallery.Name}'");
+                            }
+                            else if (action == ReviewAction.UNKNOWN)
+                            {
+                                throw new Exception(_localizer["Unknown_Review_Action"].Value);
                             }
 
-                            await _database.DeleteGalleryItem(review);
-
-                            await _audit.LogAction(User?.Identity?.Name, $"'{review.Title}' {_localizer["Audit_ItemRejectedInGallery"].Value} '{review.GalleryName}'");
+                            return Json(new { success = true, action });
                         }
-                        else if (action == ReviewAction.UNKNOWN)
-                        {
-                            throw new Exception(_localizer["Unknown_Review_Action"].Value);
-                        }
-
-                        return Json(new { success = true, action });
                     }
                     else
                     {
@@ -531,40 +524,47 @@ namespace WeddingShare.Controllers
                     var items = await _database.GetPendingGalleryItems();
                     if (items != null && items.Any())
                     {
-                        foreach (var review in items)
-                        { 
-                            var galleryDir = Path.Combine(UploadsDirectory, review.GalleryName);
-                            var reviewFile = Path.Combine(galleryDir, "Pending", review.Title);
-                            if (action == ReviewAction.APPROVED)
+                        foreach (var galleryGroup in items.GroupBy(x => x.GalleryId))
+                        {
+                            var gallery = await _database.GetGallery(galleryGroup.Key);
+                            if (gallery != null)
                             {
-                                _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(galleryDir, review.Title));
-
-                                review.State = GalleryItemState.Approved;
-                                await _database.EditGalleryItem(review);
-
-                                await _audit.LogAction(User?.Identity?.Name, _localizer["Audit_BulkApproveReviews"].Value);
-                            }
-                            else if (action == ReviewAction.REJECTED)
-                            {
-                                var retain = await _settings.GetOrDefault(Settings.Gallery.RetainRejectedItems, false);
-                                if (retain)
+                                foreach (var review in galleryGroup)
                                 {
-                                    var rejectedDir = Path.Combine(galleryDir, "Rejected");
-                                    _fileHelper.CreateDirectoryIfNotExists(rejectedDir);
-                                    _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(rejectedDir, review.Title));
-                                }
-                                else
-                                {
-                                    _fileHelper.DeleteFileIfExists(reviewFile);
-                                }
+                                    var galleryDir = Path.Combine(UploadsDirectory, gallery.Name);
+                                    var reviewFile = Path.Combine(galleryDir, "Pending", review.Title);
+                                    if (action == ReviewAction.APPROVED)
+                                    {
+                                        _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(galleryDir, review.Title));
 
-                                await _database.DeleteGalleryItem(review);
+                                        review.State = GalleryItemState.Approved;
+                                        await _database.EditGalleryItem(review);
 
-                                await _audit.LogAction(User?.Identity?.Name, _localizer["Audit_BulkRejectReviews"].Value);
-                            }
-                            else if (action == ReviewAction.UNKNOWN)
-                            {
-                                throw new Exception(_localizer["Unknown_Review_Action"].Value);
+                                        await _audit.LogAction(User?.Identity?.Name, _localizer["Audit_BulkApproveReviews"].Value);
+                                    }
+                                    else if (action == ReviewAction.REJECTED)
+                                    {
+                                        var retain = await _settings.GetOrDefault(Settings.Gallery.RetainRejectedItems, false);
+                                        if (retain)
+                                        {
+                                            var rejectedDir = Path.Combine(galleryDir, "Rejected");
+                                            _fileHelper.CreateDirectoryIfNotExists(rejectedDir);
+                                            _fileHelper.MoveFileIfExists(reviewFile, Path.Combine(rejectedDir, review.Title));
+                                        }
+                                        else
+                                        {
+                                            _fileHelper.DeleteFileIfExists(reviewFile);
+                                        }
+
+                                        await _database.DeleteGalleryItem(review);
+
+                                        await _audit.LogAction(User?.Identity?.Name, _localizer["Audit_BulkRejectReviews"].Value);
+                                    }
+                                    else if (action == ReviewAction.UNKNOWN)
+                                    {
+                                        throw new Exception(_localizer["Unknown_Review_Action"].Value);
+                                    }
+                                }
                             }
                         }
                     }
@@ -590,7 +590,7 @@ namespace WeddingShare.Controllers
                 {
                     try
                     {
-                        var check = await _database.GetGallery(model.Name);
+                        var check = await _database.GetGallery(model.Id);
                         if (check == null)
                         {
                             if (await _database.GetGalleryCount() < await _settings.GetOrDefault(Settings.Basic.MaxGalleryCount, 1000000))
@@ -634,7 +634,7 @@ namespace WeddingShare.Controllers
                 {
                     try
                     {
-                        var check = await _database.GetGallery(model.Name);
+                        var check = await _database.GetGallery(model.Id);
                         if (check == null || model.Id == check.Id)
                         {
                             var gallery = await _database.GetGallery(model.Id);
@@ -1070,7 +1070,7 @@ namespace WeddingShare.Controllers
                                 {
                                     Id = m.Key,
                                     Value = m.Value
-                                }, gallery?.Name ?? string.Empty);
+                                }, gallery?.Id);
 
                                 if (setting == null || (setting.Value ?? string.Empty) != (m.Value ?? string.Empty))
                                 {
@@ -1556,6 +1556,43 @@ namespace WeddingShare.Controllers
             {
                 return false;
             }
+        }
+
+        private async Task<List<PhotoGallery>> GetPendingReviews()
+        {
+            var galleries = new List<PhotoGallery>();
+
+            var items = !await _settings.GetOrDefault(Settings.Basic.SingleGalleryMode, false) ? await _database.GetPendingGalleryItems() : await _database.GetPendingGalleryItems(1);
+            if (items != null)
+            {
+                foreach (var galleryGroup in items.GroupBy(x => x.GalleryId))
+                {
+                    var gallery = await _database.GetGallery(galleryGroup.Key);
+                    if (gallery != null)
+                    {
+                        galleries.Add(new PhotoGallery()
+                        {
+                            GalleryId = gallery?.Id,
+                            GalleryName = gallery?.Name,
+                            Images = items?.Select(x => new PhotoGalleryImage()
+                            {
+                                Id = x.Id,
+                                GalleryId = x.GalleryId,
+                                Name = Path.GetFileName(x.Title),
+                                UploadedBy = x.UploadedBy,
+                                UploadDate = x.UploadedDate,
+                                ImagePath = $"/{Path.Combine(UploadsDirectory, gallery.Name).Remove(_hostingEnvironment.WebRootPath).Replace('\\', '/').TrimStart('/')}/Pending/{x.Title}",
+                                ThumbnailPath = $"/{Path.Combine(ThumbnailsDirectory, gallery.Name).Remove(_hostingEnvironment.WebRootPath).Replace('\\', '/').TrimStart('/')}/{Path.GetFileNameWithoutExtension(x.Title)}.webp",
+                                ThumbnailPathFallback = $"/{ThumbnailsDirectory.Remove(_hostingEnvironment.WebRootPath).Replace('\\', '/').TrimStart('/')}/{Path.GetFileNameWithoutExtension(x.Title)}.webp",
+                                MediaType = x.MediaType
+                            })?.ToList(),
+                            ItemsPerPage = int.MaxValue,
+                        });
+                    }
+                }
+            }
+
+            return galleries;
         }
     }
 }
